+++ CHANGELOG.md
@@ -7,7 +7,7 @@ The format follows [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and
 
 ## [Unreleased]
 
-## [1.3.0] - 2026-08-06
+## [1.3.0] - 2026-08-10
 
 ### Added